# build/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"ISubscription": {
				"properties": {
					"organizationId": {
						"type": "string"
					},
					"planId": {
						"type": "string"
					},
					"status": {
						"type": "string"
					},
					"startDate": {
						"type": "string",
						"format": "date-time"
					},
					"endDate": {
						"type": "string",
						"format": "date-time"
					},
					"autoRenew": {
						"type": "boolean"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"organizationId",
					"planId",
					"startDate",
					"endDate"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Subscription": {
				"properties": {
					"organizationId": {
						"type": "string"
					},
					"planId": {
						"type": "string"
					},
					"status": {
						"type": "string"
					},
					"startDate": {
						"type": "string",
						"format": "date-time"
					},
					"endDate": {
						"type": "string",
						"format": "date-time"
					},
					"autoRenew": {
						"type": "boolean"
					}
				},
				"required": [
					"organizationId",
					"planId",
					"startDate",
					"endDate"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_Subscription_": {
				"properties": {
					"organizationId": {
						"type": "string"
					},
					"planId": {
						"type": "string"
					},
					"status": {
						"type": "string"
					},
					"startDate": {
						"type": "string",
						"format": "date-time"
					},
					"endDate": {
						"type": "string",
						"format": "date-time"
					},
					"autoRenew": {
						"type": "boolean"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"ModuleType": {
				"type": "string",
				"enum": [
					"cpu",
					"memory",
					"storage",
					"network",
					"antivirus"
				]
			},
			"IPlan": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"maxMachines": {
						"type": "number",
						"format": "double"
					},
					"allowedModules": {
						"items": {
							"$ref": "#/components/schemas/ModuleType"
						},
						"type": "array"
					},
					"billingPeriod": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"isActive": {
						"type": "boolean"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"maxMachines",
					"allowedModules",
					"billingPeriod",
					"price"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Plan": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"maxMachines": {
						"type": "number",
						"format": "double"
					},
					"allowedModules": {
						"items": {
							"$ref": "#/components/schemas/ModuleType"
						},
						"type": "array"
					},
					"billingPeriod": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"isActive": {
						"type": "boolean"
					}
				},
				"required": [
					"name",
					"maxMachines",
					"allowedModules",
					"billingPeriod",
					"price"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_Plan_": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"maxMachines": {
						"type": "number",
						"format": "double"
					},
					"allowedModules": {
						"items": {
							"$ref": "#/components/schemas/ModuleType"
						},
						"type": "array"
					},
					"billingPeriod": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"isActive": {
						"type": "boolean"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"IOrganization": {
				"properties": {
					"name": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"parentId": {
						"type": "string"
					},
					"document": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"isActive": {
						"type": "boolean"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"type"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Organization": {
				"properties": {
					"name": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"parentId": {
						"type": "string"
					},
					"document": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"isActive": {
						"type": "boolean"
					}
				},
				"required": [
					"name",
					"type"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_Organization_": {
				"properties": {
					"name": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"parentId": {
						"type": "string"
					},
					"document": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"isActive": {
						"type": "boolean"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "X-Token",
				"in": "header",
				"description": "JWT token sent via X-Token header. Example: \"X-Token: {token}\""
			}
		}
	},
	"info": {
		"title": "api-tor",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/subscriptions": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "List",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ISubscription"
									},
									"type": "array"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"Subscription"
				],
				"security": [
					{
						"jwt": [
							"master"
						]
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"id": {
											"type": "string"
										}
									},
									"required": [
										"id"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"Subscription"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Subscription"
							}
						}
					}
				}
			}
		},
		"/subscriptions/organization/{organizationId}": {
			"get": {
				"operationId": "GetByOrganization",
				"responses": {
					"200": {
						"description": "Subscriptions",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ISubscription"
									},
									"type": "array"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"Subscription"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "organizationId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/subscriptions/{id}": {
			"get": {
				"operationId": "GetById",
				"responses": {
					"200": {
						"description": "Subscription",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ISubscription"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"404": {
						"description": "Not Found"
					},
					"500": {
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"Subscription"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Updated"
					},
					"500": {
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"Subscription"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_Subscription_"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Deleted"
					},
					"500": {
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"Subscription"
				],
				"security": [
					{
						"jwt": [
							"master"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/subscriptions/validate/{organizationId}/{module}": {
			"get": {
				"operationId": "ValidateSubscription",
				"responses": {
					"200": {
						"description": "Validation Result",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"reason": {
											"type": "string"
										},
										"valid": {
											"type": "boolean"
										}
									},
									"required": [
										"valid"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"Subscription"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "organizationId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "module",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/ModuleType"
						}
					}
				]
			}
		},
		"/plans": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "List",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IPlan"
									},
									"type": "array"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"Plan"
				],
				"security": [
					{
						"jwt": [
							"master"
						]
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"id": {
											"type": "string"
										}
									},
									"required": [
										"id"
									],
									"type": "object"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request"
					},
					"500": {
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"Plan"
				],
				"security": [
					{
						"jwt": [
							"master"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Plan"
							}
						}
					}
				}
			}
		},
		"/plans/{id}": {
			"get": {
				"operationId": "GetById",
				"responses": {
					"200": {
						"description": "Plan",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IPlan"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"404": {
						"description": "Not Found"
					},
					"500": {
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"Plan"
				],
				"security": [
					{
						"jwt": [
							"master"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Updated"
					},
					"500": {
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"Plan"
				],
				"security": [
					{
						"jwt": [
							"master"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_Plan_"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Deleted"
					},
					"500": {
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"Plan"
				],
				"security": [
					{
						"jwt": [
							"master"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/organizations": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "List",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IOrganization"
									},
									"type": "array"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"Organization"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"id": {
											"type": "string"
										}
									},
									"required": [
										"id"
									],
									"type": "object"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request"
					},
					"500": {
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"Organization"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Organization"
							}
						}
					}
				}
			}
		},
		"/organizations/{id}": {
			"get": {
				"operationId": "GetById",
				"responses": {
					"200": {
						"description": "Organization",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IOrganization"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"404": {
						"description": "Not Found"
					},
					"500": {
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"Organization"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Updated"
					},
					"500": {
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"Organization"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_Organization_"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Deleted"
					},
					"400": {
						"description": "Bad Request"
					},
					"500": {
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"Organization"
				],
				"security": [
					{
						"jwt": [
							"master"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/organizations/company/{companyId}/customers": {
			"get": {
				"operationId": "GetCustomersByCompany",
				"responses": {
					"200": {
						"description": "Customers",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IOrganization"
									},
									"type": "array"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"Organization"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "companyId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}